requirements/base.txt
celery==5.3.6
Django==4.2.3
django-filter==23.2
django-environ==0.10.0
drf-spectacular==0.26.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
requests-oauthlib==1.3.1
django-jalali-date==1.0.2
ipython==8.14.0
mysqlclient==2.2.0
pyOpenSSL==24.1.0
Pillow==10.1.0
requests==2.32.2
requests-oauthlib==1.3.1
